Fix macOS VST3 release compatibility

+++ src/vst3/transport_utils.rs
@@ -8,10 +8,10 @@ pub(super) fn transport_state_from_vst3_process_context(
     };
 
     let state = process_context.state;
-    let tempo_valid = (state & (ProcessContext_::StatesAndFlags_::kTempoValid as u32)) != 0;
+    let tempo_valid = (state & ProcessContext_::StatesAndFlags_::kTempoValid) != 0;
     let project_time_music_valid =
-        (state & (ProcessContext_::StatesAndFlags_::kProjectTimeMusicValid as u32)) != 0;
-    let is_playing = (state & (ProcessContext_::StatesAndFlags_::kPlaying as u32)) != 0;
+        (state & ProcessContext_::StatesAndFlags_::kProjectTimeMusicValid) != 0;
+    let is_playing = (state & ProcessContext_::StatesAndFlags_::kPlaying) != 0;
     TransportState {
         tempo_bpm: if tempo_valid {
             process_context.tempo as f32
